backend/requirements.txt
django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
python-decouple
channels
daphne
razorpay
groq